requirements.txt
nose
grpcio
grpcio-tools
googleapis-common-protos
